requirements.txt
pbr>=0.6,<1.0

Paste
PasteDeploy>=1.5.0
Routes>=1.12.3,!=2.0
amqplib>=0.6.1
anyjson>=0.3.3
argparse
Babel>=1.3
eventlet>=0.13.0,<0.16.0
greenlet>=0.3.2
httplib2>=0.7.5
requests>=1.1
iso8601>=0.1.9
jsonrpclib
Jinja2
kombu>=2.4.8
netaddr>=0.7.6
python-neutronclient>=2.3.4,<2.4
SQLAlchemy>=0.7.8,!=0.9.5,<=0.9.99
WebOb>=1.2.3
python-keystoneclient>=0.7.0,<0.12
alembic>=0.4.1
six>=1.6.0
stevedore>=0.14,<1.2
oslo.config>=1.2.0,<1.5
oslo.rootwrap<1.4

python-novaclient>=2.17.0,<2.21
taskflow>=0.1.3,<0.2
